localization added; small refactoring
